+++ packages/core/src/manglers/index.ts
@@ -1,1 +1,8 @@
+export { default as BuiltInManglers } from "./builtin";
+export { default as CssClassMangler } from "./css-classes";
+export { default as CssVariableMangler } from "./css-variables";
+export { default as HtmlAttributeMangler } from "./html-attributes";
+export { default as HtmlIdMangler } from "./html-ids";
+export { default as RecommendedManglers } from "./recommended";
+
 export type { WebManglerPlugin } from "./types";